solution-blueprints/document-summarization/app/components/requirements.txt
# Copyright © Advanced Micro Devices, Inc., or its affiliates.
#
# SPDX-License-Identifier: MIT

# Minimal, explicitly-used dependencies for `document-summarization/app/components`.

aiofiles==24.1.0
aiohttp==3.12.15
docarray==0.41.0
docx2txt==0.9
fastapi==0.116.1
httpx==0.28.1
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.4
langchain-openai==1.1.6
langchain-text-splitters==1.0.0
numpy==2.3.2
pillow==11.3.0
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
pydantic==2.11.7
pypdf==5.4.0
python-multipart==0.0.20
requests==2.32.5
shortuuid==1.0.13
starlette==0.47.2
tiktoken==0.12.0
transformers==4.57.0
urllib3==2.5.0
uvicorn==0.35.0
